Add SEO expandable section component and use on home page
